Add bsb00046285
Note that the IIIF lists the license as http://www.europeana.eu/rights/rr-f/
which gives no details beyond "in copyright, see license". However
the web interface for the book at http://daten.digitale-sammlungen.de/~db/0004/bsb00046285/images/index.html
states that the license is http://creativecommons.org/licenses/by-nc-sa/4.0/
so that's what we're assuming is correct, and how it is listed here.