Add RDF metadata links to all book metadata, and fix one license link
I had incorrectly put the CC-Zero license link in one metadata.txt
file, as that is what is stated in the RDF for all of the books.
However I think this refers to the copyright status of the books
themselves, not the page images, which are claimed to be CC-BY-NC-SA
by BSB, presumably under the Sweat of the Brow doctrine.